Swap to Monorepo structure with uv workspaces
Swap out to have 4 packages:
- divbase-api (FastAPI backend)
- divbase-worker (Celery worker)
- divbase-cli (CLI tool)
- divbase-lib (shared library code)
