+++ docker/fastapi.dockerfile
@@ -5,13 +5,21 @@ WORKDIR /app
 # Install curl for healthcheck
 RUN apk add --no-cache curl gcc g++ musl-dev
 
-# copy readme to avoid pip complaining about missing files
-COPY pyproject.toml README.md ./
-
 RUN pip install --upgrade pip 
 
-COPY src/ ./src/
-RUN pip install -e .
+# Copy workspace files, TODO - do I need outer pyproject.toml?
+COPY pyproject.toml ./
+COPY README.md ./
+
+# Copy all package sources
+COPY packages/divbase-lib/ ./packages/divbase-lib/
+COPY packages/divbase-worker/ ./packages/divbase-worker/
+COPY packages/divbase-api/ ./packages/divbase-api/
+
+# Install packages in dependency order
+RUN pip install -e ./packages/divbase-lib/
+RUN pip install -e ./packages/divbase-worker/
+RUN pip install -e ./packages/divbase-api/
 
 # host needs to be set to 0.0.0.0 to be accessible from outside the container
-CMD ["fastapi", "run", "--host", "0.0.0.0", "/app/src/divbase_tools/divbase_api.py"]
+CMD ["fastapi", "run", "--host", "0.0.0.0", "/app/packages/divbase-api/src/divbase_api/divbase_api.py"]
