+++ docker/flower.dockerfile
@@ -5,12 +5,19 @@ WORKDIR /app
 # Install curl for healthcheck
 RUN apk add --no-cache curl gcc g++ musl-dev
 
-# copy readme to avoid pip complaining about missing files
-COPY pyproject.toml README.md ./
 
 RUN pip install --upgrade pip 
 
-COPY src/ ./src/
-RUN pip install -e .
+# Copy workspace files, TODO - do I need outer pyproject.toml?
+COPY pyproject.toml ./
+COPY README.md ./
 
-ENTRYPOINT ["celery", "-A", "divbase_tools.tasks"]
+# Copy all package sources
+COPY packages/divbase-lib/ ./packages/divbase-lib/
+COPY packages/divbase-worker/ ./packages/divbase-worker/
+
+# Install packages in dependency order
+RUN pip install -e ./packages/divbase-lib/
+RUN pip install -e ./packages/divbase-worker/
+
+ENTRYPOINT ["celery", "-A", "divbase_worker.tasks"]
