+++ packages/divbase-api/pyproject.toml
@@ -1,0 +1,23 @@
+[project]
+name = "divbase-api"
+version = "0.1.0"
+readme = "../../README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "fastapi[standard]>=0.115.12",
+    "uvicorn>=0.34.3",
+    "httpx>=0.28.1",
+    "divbase-lib",
+    "divbase-worker",
+]
+
+[build-system]
+requires = ["hatchling >= 1.26"]
+build-backend = "hatchling.build"
+
+[tool.uv.sources]
+divbase-lib = { workspace = true }
+divbase-worker = { workspace = true }
+
+[project.scripts]
+divbase-api = "divbase_api.divbase_api:main"
