+++ packages/divbase-api/src/divbase_api/__init__.py
@@ -1,0 +1,3 @@
+# File: /divbase/packages/divbase-api/src/divbase_api/__init__.py
+
+"""This package provides the API functionalities for the divbase project."""
