+++ packages/divbase-cli/src/divbase_cli/__init__.py
@@ -1,0 +1,3 @@
+# File: /divbase/packages/divbase-cli/src/divbase_cli/__init__.py
+
+"""This file initializes the divbase_cli package."""
