+++ packages/divbase-cli/src/divbase_cli/bucket_versioning.py
@@ -9,14 +9,13 @@ from dataclasses import dataclass, field
 from datetime import datetime, timezone
 
 import yaml
-
-from divbase_tools.exceptions import (
+from divbase_lib.exceptions import (
     BucketVersionAlreadyExistsError,
     BucketVersioningFileAlreadyExistsError,
     BucketVersionNotFoundError,
     ObjectDoesNotExistError,
 )
-from divbase_tools.s3_client import S3FileManager
+from divbase_lib.s3_client import S3FileManager
 
 VERSION_FILE_NAME = ".bucket_versions.yaml"
