+++ packages/divbase-cli/src/divbase_cli/cli_commands/__init__.py
@@ -1,0 +1,4 @@
+# File: /divbase/divbase/packages/divbase-cli/src/divbase_cli/commands/__init__.py
+"""
+This file initializes the commands subpackage for the divbase-cli package.
+"""
