+++ packages/divbase-cli/src/divbase_cli/user_config.py
@@ -12,8 +12,7 @@ from dataclasses import dataclass, field
 from pathlib import Path
 
 import yaml
-
-from divbase_tools.exceptions import ProjectNotInConfigError
+from divbase_lib.exceptions import ProjectNotInConfigError
 
 
 @dataclass
