+++ packages/divbase-lib/pyproject.toml
@@ -1,0 +1,14 @@
+[project]
+name = "divbase-lib"
+version = "0.1.0"
+readme = "../../README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "pyyaml>=6.0.2",
+    "boto3>=1.38.15",
+    "pandas>=2.2.3",
+]
+
+[build-system]
+requires = ["hatchling >= 1.26"]
+build-backend = "hatchling.build"
