+++ packages/divbase-lib/src/divbase_lib/__init__.py
@@ -1,0 +1,3 @@
+# File: /divbase/packages/divbase-lib/src/divbase_lib/__init__.py
+
+"""This package provides core functionalities for the divbase library."""
