+++ packages/divbase-worker/pyproject.toml
@@ -1,0 +1,18 @@
+[project]
+name = "divbase-worker"
+version = "0.1.0"
+readme = "../../README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "celery>=5.5.3",
+    "redis>=4.5.0",
+    "httpx>=0.28.1",
+    "divbase-lib",
+]
+
+[build-system]
+requires = ["hatchling >= 1.26"]
+build-backend = "hatchling.build"
+
+[tool.uv.sources]
+divbase-lib = { workspace = true }
