+++ packages/divbase-worker/src/divbase_worker/__init__.py
@@ -1,0 +1,3 @@
+# File: /divbase/divbase/packages/divbase-worker/src/divbase_worker/__init__.py
+
+"""This file initializes the divbase_worker package."""
